requirements.txt
bm25s[full]
PyStemmer
konlpy
kiwipiepy